Backend/requirements.txt
Flask==1.1.2
Flask-cors>=3.0.7
Flask-Script>=2.0.6
Flask-Injector>=0.12.0
Flask-SQLAlchemy>=2.3.2
pandas>=0.24.1
pyPika>=0.34.0
Werkzeug==1.0.1
cx_Oracle